requirements.txt
fastapi==0.102.0        # latest FastAPI compatible with Python 3.13
uvicorn==0.23.1
sqlmodel==0.0.25
pydantic==2.11.9
PyYAML==6.0.2

